% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_censor_mark.R
\name{add_censor_mark}
\alias{add_censor_mark}
\title{Add Censor Marking}
\usage{
add_censor_mark(shape = 3, size = 2, ...)
}
\arguments{
\item{shape, size, ...}{arguments passed to passed to
\code{ggplot2::geom_point(shape, size, ...)}}
}
\value{
a ggplot
}
\description{
Add Censor Marking
}
\examples{
survfit2(Surv(time, status) ~ 1, data = df_lung) \%>\%
  ggsurvfit() +
  add_confidence_interval() +
  add_censor_mark()
}
