% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_legend_title.R
\name{add_legend_title}
\alias{add_legend_title}
\title{Add Legend Title}
\usage{
add_legend_title(title = NULL)
}
\arguments{
\item{title}{a string to override the default legend title. Default is \code{NULL}}
}
\value{
a ggplot2 figure
}
\description{
Add a default or custom title to the figure legend.
}
\examples{
survfit2(Surv(time, status) ~ surg, data = df_colon) \%>\%
  ggsurvfit() +
  add_legend_title()
}
\seealso{
Visit the \href{https://www.danieldsjoberg.com/ggsurvfit/articles/gallery.html}{gallery} for examples modifying the default figures
}
