% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_quantile.R
\name{add_quantile}
\alias{add_quantile}
\title{Add Quantile}
\usage{
add_quantile(y_value = 0.5, linetype = 2, ...)
}
\arguments{
\item{y_value}{numeric value where the line segment will be drawn. Default is \code{0.5}}

\item{linetype}{Default is \code{"dashed"}. Argument is passed to \code{ggplot2::geom_segment()}}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}
}
\value{
a ggplot
}
\description{
Add quantile information to the plot.
}
\examples{
survfit2(Surv(time, status) ~ sex, data = df_lung) \%>\%
  ggsurvfit() +
  add_quantile(linetype = 2)

survfit2(Surv(time, status) ~ 1, data = df_lung) \%>\%
  ggsurvfit() +
  add_quantile(linetype = 2) +
  add_quantile(y_value = 0.9, linetype = 3)
}
