% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_risktable.R
\name{add_risktable}
\alias{add_risktable}
\title{Add risk tables}
\usage{
add_risktable(
  times = NULL,
  risktable_stats = c("n.risk", "cum.event"),
  risktable_group = c("strata", "risktable_stats"),
  risktable_height = 0.14,
  stats_label = NULL,
  combine_groups = FALSE,
  theme = theme_ggsurvfit_risktable()
)
}
\arguments{
\item{times}{numeric vector of times where risk table values will be placed.
Default are the times shown on the x-axis.}

\item{risktable_stats}{character vector of statistics to show in the risk table.
Must be one or more of \code{c("n.risk", "cum.event", "cum.censor", "n.event", "n.censor")}.
Default is \code{c("n.risk", "cum.event")}.
\itemize{
\item \code{"n.risk"} Number of patients at risk
\item \code{"cum.event"} Cumulative number of observed events
\item \code{"cum.censor"} Cumulative number of censored observations
\item \code{"n.event"} Number of events in time interval
\item \code{"n.censor"} Number of censored observations in time interval
}}

\item{risktable_group}{String indicating the grouping variable for the risk tables.
Default is \code{"strata"}.
\itemize{
\item \code{"strata"} groups the risk tables per stratum when present.
\item \code{"risktable_stats"} groups the risk tables per risktable_stats.
}}

\item{risktable_height}{A numeric value between 0 and 1 indicates the height used by the table versus the height
used by the plot, as described in \code{patchwork::wrap_plots(heights=)}. The default is 0.14.}

\item{stats_label}{named vector or list of custom labels. Names are the statistics
from \verb{risktable_stats=} and the value is the custom label.}

\item{combine_groups}{logical indicating whether to combine the statistics
in the risk table across groups. Default is \code{FALSE}}

\item{theme}{A risktable theme typically returned from \code{theme_ggsurvfit_risktable()}}
}
\description{
Add risk tables
}
\examples{
library(ggplot2)

p <-
  df_lung \%>\%
  survfit2(Surv(time, status) ~ sex, data = .) \%>\%
  ggsurvfit() +
  add_censor_mark() +
  add_confidence_interval()

p + add_risktable()

p +
  add_risktable(
    risktable_stats = c("n.risk", "cum.event"),
    stats_label = list(
      cum.event = "Cumulative Observed Events",
      n.risk = "Number at Risk"
    ),
    risktable_group = "strata",
  )

p +
  add_risktable(
    risktable_stats = c("n.risk", "cum.event"),
    combine_groups = TRUE
  )
}
