% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_risktable_strata_symbol.R
\name{add_risktable_strata_symbol}
\alias{add_risktable_strata_symbol}
\title{Use Symbol for Strata in Risk Table}
\usage{
add_risktable_strata_symbol(
  symbol = NULL,
  size = 15,
  face = "bold",
  vjust = 0.3,
  ...
)
}
\arguments{
\item{symbol}{UTF-8 code of shape to replace strata level with.
Default is a rectangle (\code{"\\U25AC"}).
Other common options are circle (\code{"\\U25CF"}) and diamond (\code{"\\U25C6"})}

\item{size, face, vjust, ...}{arguments passed to \code{ggtext::element_markdown()}
to style the symbol}
}
\value{
a ggplot
}
\description{
Use Symbol for Strata in Risk Table
}
\examples{
p <-
survfit2(Surv(time, status) ~ sex, data = df_lung) \%>\%
  ggsurvfit(size = 1) +
  add_confidence_interval() +
  add_risktable(risktable_group = "risktable_stats")

 p + add_risktable_strata_symbol()
 p + add_risktable_strata_symbol(symbol = "\U25CF")
}
