% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{survfit_pvalue}
\title{Deprecated functions}
\usage{
survfit_pvalue(...)
}
\description{
Some functions have been deprecated and are no longer being actively
supported.
}
\keyword{internal}
