% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_colon}
\alias{df_colon}
\title{Formatted Copy of \code{survival::colon}}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 929 rows and 14 columns.
}
\usage{
df_colon
}
\description{
This is a copy of the colon data set exported by the survival
package. This data set, however, has column labels,
numeric categorical variables are now factors with assigned levels,
and we only include the recurrence outcome.
}
\keyword{datasets}
