% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_lung}
\alias{df_lung}
\title{Formatted Copy of \code{survival::lung}}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 228 rows and 10 columns.
}
\usage{
df_lung
}
\description{
This is a copy of the lung data set exported by the survival
package. This data set, however, has column labels and
numeric categorical variables are now factors with assigned levels.
}
\keyword{datasets}
