% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_p.R
\name{format_p}
\alias{format_p}
\title{Format p-value}
\usage{
format_p(x, digits = 1)
}
\arguments{
\item{x}{numeric vector of p-values}

\item{digits}{number of digits large p-values will be rounded to.
Default is 2, and must be one of 1, 2, or 3.}
}
\value{
a string
}
\description{
Round and format p-values
}
\examples{
p_vec <- c(0.00001, 0.01111, 0.0500000, 0.15, 0.99999)
format_p(p_vec)
format_p(p_vec, 2)
format_p(p_vec, 3)
}
