% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcuminc.R
\name{ggcuminc}
\alias{ggcuminc}
\title{Plot Cumulative Incidence}
\usage{
ggcuminc(x, outcome = NULL, theme = theme_ggsurvfit_default(), ...)
}
\arguments{
\item{x}{a 'cuminc' object created with \code{tidycmprsk::cuminc()}}

\item{outcome}{string indicating which outcome to include in plot.
Default is to include the first competing event.}

\item{theme}{a theme . Default is \code{theme_ggsurvfit_default()}}

\item{...}{arguments passed to \code{ggplot2::geom_step(...)}, e.g. \code{size = 2}}
}
\value{
a ggplot
}
\description{
Plot Cumulative Incidence
}
\examples{
library(tidycmprsk)

cuminc(Surv(ttdeath, death_cr) ~ trt, trial) \%>\%
  ggcuminc(outcome = "death from cancer") +
  add_confidence_interval() +
  add_risktable()
}
