% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsurvfit_arithmetic.R
\name{ggsurvfit_arithmetic}
\alias{ggsurvfit_arithmetic}
\alias{|.ggsurvfit_build}
\alias{|.ggsurvfit}
\alias{|.ggcuminc}
\alias{-.ggsurvfit_build}
\alias{-.ggsurvfit}
\alias{-.ggcuminc}
\alias{/.ggsurvfit_build}
\alias{/.ggsurvfit}
\alias{/.ggcuminc}
\title{Plot Arithmetic}
\usage{
\method{|}{ggsurvfit_build}(e1, e2)

\method{|}{ggsurvfit}(e1, e2)

\method{|}{ggcuminc}(e1, e2)

\method{-}{ggsurvfit_build}(e1, e2)

\method{-}{ggsurvfit}(e1, e2)

\method{-}{ggcuminc}(e1, e2)

\method{/}{ggsurvfit_build}(e1, e2)

\method{/}{ggsurvfit}(e1, e2)

\method{/}{ggcuminc}(e1, e2)
}
\arguments{
\item{e1}{A \code{ggsurvfit}, \code{ggcuminc}, or \code{ggsurvfit_built} object}

\item{e2}{A \code{ggsurvfit}, \code{ggcuminc},\code{ggsurvfit_built}, \code{ggplot} or \code{patchwork}
object in case of \code{/}, or a \code{gg} object such as a geom or
theme specification in case of \code{*} and \code{&}}
}
\value{
A \code{patchwork} object
}
\description{
Plot arithmetic for 'ggsurvfit' figures, built to mimic the
\href{https://patchwork.data-imaginist.com/reference/plot_arithmetic.html}{patchwork plot arithmetic}.
Methods for \code{|}, \code{-}, and \code{/} have been implemented.

Use these functions to patch together figures created with the ggsurvfit package.
If you need patch together 'ggsurvfit' plots with other plots, see details below.

If you experience an error using these operators, \strong{read the details section below}.
}
\section{Details}{


A few details on ggsurvfit plot arithmetic:
\enumerate{
\item These functions are meant to only work with figures created with the ggsurvfit package,
and only two plots at a time. This means you can do this: \code{p1 | p2}. But
not this: \code{p1 | p2 | p3}. The reason for this is that these methods are
meant to be operated on figures created with ggsurvfit. After the first
\code{p1 | p2} is executed, the returned plot is no longer of class 'ggsurvfit',
and therefore the remaining operation of \code{. | p3} cannot execute properly.
See below for details on operating on 3 or more figures.
\item Not every combination of operators is supported by the ggsurvfit
implementation of the plot operators. To utilize the full suite supported
by patchwork, see the note below.
\item If you need to patch together 'ggsurvfit' figures with non-ggsurvfit
figures OR you need a feature supported in patchwork that is not supported
by the ggsurvfit implementation, you must utilize the arithmetic
operators directly from the \code{patchwork} package.
To use directly use the \code{patchwork} operators with a figure created
with \code{ggsurvfit}, it must be wrapped in \code{patchwork::wrap_elements(ggsurvfit_build(p))}.

\if{html}{\out{<div class="sourceCode r">}}\preformatted{patchwork::wrap_elements(ggsurvfit_build(p)) + patchwork::wrap_elements(ggsurvfit_build(p))
}\if{html}{\out{</div>}}
}
}

\examples{
p <-
  survfit2(Surv(time, status) ~ sex, data = df_lung) \%>\%
  ggsurvfit() +
  add_risktable()

p1 <- p + ggtitle("Plot 1")
p2 <- p + ggtitle("Plot 2")

p1 | p2

p1 - p2
}
