% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsurvfit_options.R
\name{ggsurvfit_options}
\alias{ggsurvfit_options}
\title{Global Options}
\description{
By default, \code{ggsurvfit()} and \code{ggcuminc()} uses the color aesthetic to draw
curves stratified by treatment group. Moreover, in \code{ggcuminc()} when multiple
outcomes are plotted on the same figure the linetype aesthetic is used to
distinguish the curves among the various outcomes.

It is, however, sometimes desirable to use the linetype to stratify
by treatment group and color by outcome. To obtain these figures, set the
\code{options("ggsurvfit.switch-color-linetype" = TRUE)} option.
}
\examples{
options("ggsurvfit.switch-color-linetype" = TRUE)
library(tidycmprsk)

cuminc(Surv(ttdeath, death_cr) ~ trt, trial) \%>\%
  ggcuminc(outcome = "death from cancer") +
  add_risktable()

cuminc(Surv(ttdeath, death_cr) ~ 1, trial) \%>\%
  ggcuminc(outcome = c("death from cancer", "death other causes")) +
  add_risktable()

# reset option
options("ggsurvfit.switch-color-linetype" = NULL)
}
