% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print_ggsurvfit}
\alias{print_ggsurvfit}
\alias{print.ggsurvfit}
\alias{print.ggcuminc}
\title{Print ggsurvfit object}
\usage{
\method{print}{ggsurvfit}(x, ...)

\method{print}{ggcuminc}(x, ...)
}
\arguments{
\item{x}{an object of class 'ggsurvfit' or 'ggcuminc'}

\item{...}{These dots are for future extensions and must be empty.}
}
\value{
a printed ggplot2 figure
}
\description{
Print ggsurvfit object
}
\examples{
print(survfit2(Surv(time, status) ~ surg, data = df_colon))
}
\keyword{internal}
