% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{survfit}
\alias{Surv}
\alias{\%>\%}
\alias{grid.draw}
\alias{\%+\%}
\alias{ggplot_add}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr:\%>\%]{\%>\%}}}

  \item{ggplot2}{\code{\link[ggplot2:\%+\%]{\%+\%}}, \code{\link[ggplot2:ggplot_add]{ggplot_add()}}}

  \item{grid}{\code{\link[grid:grid.draw]{grid.draw()}}}

  \item{survival}{\code{\link[survival:Surv]{Surv()}}, \code{\link[survival:survfit]{survfit()}}}
}}

