% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_continuous_padding.R
\name{scale_continuous_reduce_pad}
\alias{scale_continuous_reduce_pad}
\title{Reduce Scale Padding}
\usage{
scale_continuous_reduce_pad(x_scales = list(), y_scales = list())
}
\arguments{
\item{x_scales}{a named list of arguments that will be passed to \code{ggplot2::scale_x_continuous()}.}

\item{y_scales}{a named list of arguments that will be passed to \code{ggplot2::scale_y_continuous()}.}
}
\value{
a ggplot2 figure
}
\description{
I have a hard time remembering the syntax to reduce horizontal and vertical
padding on a ggplot2 figure. Hence, a function with default reduced padding.

If you use this function, you \strong{must} include all scale specifications
that would appear in \code{scale_x_continuous()} or \code{scale_y_continuous()}.
}
\examples{
ggsurvfit <-
  survfit2(Surv(time, status) ~ surg, data = df_colon) \%>\%
  ggsurvfit(size = 1) +
  add_confidence_interval()

# use the function defaults
ggsurvfit + scale_continuous_reduce_pad()

# specify additional scales
ggsurvfit +
  scale_continuous_reduce_pad(
    x_scales = list(breaks = 0:9),
    y_scales = list(label = scales::percent, limits = c(0, 1))
  )
}
