% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survdiff2.R
\name{survdiff2}
\alias{survdiff2}
\alias{logrank}
\alias{p_logrank}
\title{Calculate p-value}
\usage{
survdiff2(x, pvalue_fun = NULL, prepend_p = TRUE, ...)

logrank(x, pvalue_fun = NULL)

p_logrank(x, pvalue_fun = NULL)
}
\arguments{
\item{x}{a 'survfit2' object}

\item{pvalue_fun}{function to round and style p-value with}

\item{prepend_p}{prepend \code{"p="} to formatted p-value}

\item{...}{
  Arguments passed on to \code{\link[survival:survdiff]{survival::survdiff}}
  \describe{
    \item{\code{subset}}{
expression indicating which subset of the rows of data should be used in
the fit.  This can be a logical vector (which is replicated to have
length equal to the number of observations), a numeric vector indicating
which observation numbers are to be included (or excluded if negative),
or a character vector of row names to be included.  All observations are
included by default.
}
    \item{\code{na.action}}{
a missing-data filter function.  This is applied to the \code{model.frame} after any
subset argument has been used.  Default is \code{options()$na.action}.
}
    \item{\code{rho}}{
a scalar parameter that controls the type of test.
}
    \item{\code{timefix}}{process times through the \code{aeqSurv} function to
  eliminate potential roundoff issues.}
  }}
}
\value{
a string
}
\description{
VERY EXPERIMENTAL! Will likely change or perhaps even disappear.
}
\examples{
# add exmaple
}
