% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survfit2_p.R
\name{survfit2_hr}
\alias{survfit2_hr}
\title{Calculate Hazard ratio}
\usage{
survfit2_hr(
  x,
  conf.int = FALSE,
  conf.level = 0.95,
  exponentiate = FALSE,
  rho = 0
)
}
\arguments{
\item{x}{A \link{survfit2} object}

\item{conf.int}{Logical indicating whether or not to include a confidence
interval in the tidied output. Defaults to \code{FALSE}.}

\item{conf.level}{The confidence level to use for the confidence interval
if \code{conf.int = TRUE}. Must be strictly greater than 0 and less than 1.
Defaults to 0.95, which corresponds to a 95 percent confidence interval.}

\item{exponentiate}{Logical indicating whether or not to exponentiate the
the coefficient estimates. This is typical for logistic and multinomial
regressions, but a bad idea if there is no log or logit link. Defaults
to \code{FALSE}.}

\item{rho}{argument passed to  \code{survival::survdiff(rho=)}}
}
\value{
An atomic numeric.
}
\description{
Calculate Hazard ratio
}
\examples{
sf <- survfit2(Surv(time, status) ~ sex, data = df_lung)
survfit2_hr(sf)
survfit2_hr(sf, conf.int = TRUE)
survfit2_hr(sf, conf.int = TRUE, exponentiate = TRUE)
}
