% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survfit2_p.R
\name{survfit2_p}
\alias{survfit2_p}
\title{Calculate p-value}
\usage{
survfit2_p(x, format = TRUE, prepend_p = TRUE, rho = 0)
}
\arguments{
\item{x}{A \link{survfit2} object}

\item{format}{A bool or a function (formula would be okay) to round and style
p-value with. If \code{FALSE}, will return pvalue directly, if \code{TRUE},
\link{survfit2_p} will be used to format the pvalue.}

\item{prepend_p}{A boolean value indicating whether a \code{"p="} should be
prepended to the formatted p-value. Only used when \code{format} is \code{TRUE} or a
function.}

\item{rho}{argument passed to  \code{survival::survdiff(rho=)}}
}
\value{
A number of pvalue if format is \code{FALSE}, otherwise, a string.
}
\description{
The function \code{survfit2_p()} wraps \code{survival::survdiff()} and returns
a formatted p-value.
}
\examples{
sf <- survfit2(Surv(time, status) ~ sex, data = df_lung)

sf \%>\%
  ggsurvfit() +
  add_confidence_interval() +
  add_risktable() +
  scale_ggsurvfit() +
  labs(caption = glue::glue("Log-rank {survfit2_p(sf)}"))

sf \%>\%
  ggsurvfit() +
  add_confidence_interval() +
  add_risktable() +
  scale_ggsurvfit() +
  annotate("text", x = 2, y = 0.05, label = glue::glue("{survfit2_p(sf)}"))
}
