% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survfit_pvalue.R
\name{survfit_pvalue}
\alias{survfit_pvalue}
\title{Compare KM curves}
\usage{
survfit_pvalue(x, test = c("logrank", "wilcoxon", "tarone_ware"), rho = NULL)
}
\arguments{
\item{x}{an object of class 'survfit'}

\item{test}{one of \code{c("logrank", "wilcoxon", "tarone_ware")}}

\item{rho}{Default is NULL. When not NULL, it's passed to \code{survdiff(rho=)} and
the \verb{test=} argument is ignored.}
}
\value{
a p-value
}
\description{
Returns p-values calculated with \code{survival::survdiff()}
}
\examples{
survfit2(Surv(time, status) ~ sex, data = df_lung) \%>\%
  survfit_pvalue(test = "logrank")
}
