% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ggsurvfit.R
\name{theme_ggsurvfit}
\alias{theme_ggsurvfit}
\alias{theme_ggsurvfit_default}
\alias{theme_ggsurvfit_KMunicate}
\title{Survfit Plot Themes}
\usage{
theme_ggsurvfit_default()

theme_ggsurvfit_KMunicate()
}
\value{
a ggplot2 theme
}
\description{
Returns ggplot list of calls defining a theme.
}
\examples{
survfit2(Surv(time, status) ~ sex, data = df_lung) \%>\%
  ggsurvfit(theme = theme_ggsurvfit_default())
}
