% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ggsurvfit_risktable.R
\name{theme_ggsurvfit_risktable}
\alias{theme_ggsurvfit_risktable}
\title{Risk Table Themes}
\usage{
theme_ggsurvfit_risktable(theme = c("default", "boxed", "none"))
}
\arguments{
\item{theme}{must be one of \code{c("default", "boxed", "none")}
\itemize{
\item \code{"default"} a clean, simple default theme
\item \code{"boxed"} the default theme with box around risk table
\item \code{"none"} no styling applied to risk table; equivalent to using \code{theme = NULL}
}}
}
\description{
Returns ggplot list of calls defining a theme meant to be applied to a risk table.
}
\examples{
p <- survfit2(Surv(time, status) ~ 1, data = df_lung) \%>\% ggsurvfit()

# default ------------------------------------
p + add_risktable(theme = theme_ggsurvfit_risktable("default"))

# boxed --------------------------------------
p + add_risktable(theme = theme_ggsurvfit_risktable("boxed"))

# none ---------------------------------------
p + add_risktable(theme = theme_ggsurvfit_risktable("none"))
}
