% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ggsurvfit_survfit.R
\name{theme_ggsurvfit_survfit}
\alias{theme_ggsurvfit_survfit}
\title{Survfit Plot Themes}
\usage{
theme_ggsurvfit_survfit(theme = c("default", "none"))
}
\arguments{
\item{theme}{must be one of \code{c("none")}
\itemize{
\item \code{"none"} no styling applied to risk table
}}
}
\description{
Returns ggplot list of calls defining a theme.
}
