% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_censor_mark.R
\name{add_censor_mark}
\alias{add_censor_mark}
\title{Add Censor Marking}
\usage{
add_censor_mark(...)
}
\arguments{
\item{...}{arguments passed to passed to
\code{ggplot2::geom_point(...)} with defaults \code{shape = 3} and \code{size = 2}}
}
\value{
a ggplot2 figure
}
\description{
Add a marking on the figure to represent the time an observations was
censored.
}
\examples{
survfit2(Surv(time, status) ~ 1, data = df_lung) \%>\%
  ggsurvfit() +
  add_confidence_interval() +
  add_censor_mark()
}
\seealso{
Visit the \href{https://www.danieldsjoberg.com/ggsurvfit/articles/gallery.html}{gallery} for examples modifying the default figures
}
