% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_pvalue.R
\name{add_pvalue}
\alias{add_pvalue}
\alias{add_pvalue_caption}
\alias{add_pvalue_annotation}
\title{Add p-value}
\usage{
add_pvalue_caption(
  caption = "{p.value}",
  prepend_p = TRUE,
  pvalue_fun = format_p,
  rho = 0
)

add_pvalue_annotation(
  caption = "{p.value}",
  prepend_p = TRUE,
  pvalue_fun = format_p,
  rho = 0,
  ...
)
}
\arguments{
\item{caption}{string to be placed as the caption/annotation. String will
be processed with \code{glue::glue()}, and the default is "{p.value}"}

\item{prepend_p}{prepend \code{"p="} to formatted p-value}

\item{pvalue_fun}{function to round and style p-value with}

\item{rho}{argument passed to  \code{survival::survdiff(rho=)}}

\item{...}{arguments passed to \code{ggplot2::annotate()}}
}
\value{
a ggplot2 figure
}
\description{
Add p-value
}
\examples{
survfit2(Surv(time, status) ~ surg, df_colon) \%>\%
  ggsurvfit() +
  add_pvalue_caption()
}
\seealso{
\code{survfit_p()}
}
