% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_pvalue.R
\name{add_pvalue}
\alias{add_pvalue}
\title{Add p-value}
\usage{
add_pvalue(
  location = c("caption", "annotation"),
  caption = "{p.value}",
  prepend_p = TRUE,
  pvalue_fun = format_p,
  rho = 0
)
}
\arguments{
\item{caption}{string to be placed as the caption/annotation. String will
be processed with \code{glue::glue()}, and the default is "{p.value}"}

\item{prepend_p}{prepend \code{"p="} to formatted p-value}

\item{pvalue_fun}{function to round and style p-value with}

\item{rho}{argument passed to  \code{survival::survdiff(rho=)}}

\item{...}{arguments passed to \code{ggplot2::annotate()}}
}
\value{
a ggplot2 figure
}
\description{
\itemize{
\item \code{add_pvalue("caption")}: Add a p-value to the figure via \code{ggplot2::labs(caption=)}
\item \code{add_pvalue("annotation")}: Add a p-value text annotation via \code{ggplot2::annotation("text")}
}

P-values are calculated with \code{survival::survdiff()}.
Examples of custom placement located in the help file for \code{survfit_p()}.
}
\examples{
survfit2(Surv(time, status) ~ surg, df_colon) \%>\%
  ggsurvfit() +
  add_pvalue()

survfit2(Surv(time, status) ~ surg, df_colon) \%>\%
  ggsurvfit() +
  add_pvalue("annotation", x = 1, y = 0.4)
}
\seealso{
\code{survfit_p()}
}
