% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsurvfit_arithmetic.R
\name{arithmetic_prep_ggsurvfit}
\alias{arithmetic_prep_ggsurvfit}
\alias{arithmetic_prep.ggsurvfit}
\alias{arithmetic_prep.ggcuminc}
\alias{arithmetic_prep.ggsurvfit_build}
\title{Patchwork Arithmetic Prep}
\usage{
\method{arithmetic_prep}{ggsurvfit}(x, ...)

\method{arithmetic_prep}{ggcuminc}(x, ...)

\method{arithmetic_prep}{ggsurvfit_build}(x, ...)
}
\arguments{
\item{x}{a plot object}

\item{...}{passed to other methods}
}
\description{
Patchwork Arithmetic Prep
}
\examples{
p <-
  survfit2(Surv(time, status) ~ sex, data = df_lung) \%>\%
  ggsurvfit() +
  add_risktable()

p | p | p
}
\keyword{internal}
