% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsurvfit-package.R
\docType{package}
\name{ggsurvfit-package}
\alias{ggsurvfit-package}
\alias{_PACKAGE}
\title{ggsurvfit: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ddsjoberg/ggsurvfit}
  \item \url{http://www.danieldsjoberg.com/ggsurvfit/}
  \item Report bugs at \url{https://github.com/ddsjoberg/ggsurvfit/issues}
}

}
\author{
\strong{Maintainer}: Daniel D. Sjoberg \email{danield.sjoberg@gmail.com} (\href{https://orcid.org/0000-0003-0862-2018}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Mark Baillie \email{bailliem@gmail.com}
}

}
\keyword{internal}
