% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsurvfit.R
\name{ggsurvfit}
\alias{ggsurvfit}
\title{Plot Survival Probability}
\usage{
ggsurvfit(x, type = "survival", theme = theme_ggsurvfit_survfit(), ...)
}
\arguments{
\item{x}{a 'survfit' object}

\item{type}{type of statistic to report. Default is \code{"survival"}.
Must be one of the following:\tabular{ll}{
   type \tab transformation \cr
   \code{"survival"} \tab \code{x} \cr
   \code{"risk"} \tab \code{1 - x} \cr
   \code{"cumhaz"} \tab \code{-log(x)} \cr
}}

\item{theme}{a survfit theme typically returned from \code{theme_ggsurvfit_survfit()}}

\item{...}{arguments passed to \code{ggplot2::geom_step(...)}, e.g. \code{size = 2}}
}
\value{
a ggplot
}
\description{
Plot Survival Probability
}
\section{Details}{


This function creates a ggplot figure from the 'survfit' object.
To better understand how to modify the figure, review the simplified
code used internally:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{survfit2(Surv(time, status) ~ sex, data = df_lung) \%>\%
  tidy_survfit() \%>\%
  ggplot(aes(x = time, y = estimate, y
             min = conf.low, ymax = conf.low,
             color = strata, fill = strata)) +
  geom_step()
}\if{html}{\out{</div>}}
}

\examples{
survfit2(Surv(time, status) ~ sex, data = df_lung) \%>\%
  ggsurvfit()

survfit2(Surv(time, status) ~ 1, data = df_lung) \%>\%
  ggsurvfit() +
  add_censor_mark() +
  add_confidence_interval() +
  add_quantile() +
  add_risktable()
}
