% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print_ggsurvfit}
\alias{print_ggsurvfit}
\alias{print.ggsurvfit}
\alias{knit_print.ggsurvfit}
\title{Print ggsurvfit object}
\usage{
\method{print}{ggsurvfit}(x, ...)

knit_print.ggsurvfit(x, ...)
}
\arguments{
\item{x}{an object of class 'ggsurvfit'}

\item{...}{These dots are for future extensions and must be empty.}
}
\value{
a printed ggplot
}
\description{
Print ggsurvfit object
}
\examples{
# add example
}
\keyword{internal}
