% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{survfit}
\alias{Surv}
\alias{\%>\%}
\alias{\%+\%}
\alias{ggplot_add}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr:reexports]{\%>\%}}}

  \item{ggplot2}{\code{\link[ggplot2:gg-add]{\%+\%}}, \code{\link[ggplot2]{ggplot_add}}}

  \item{survival}{\code{\link[survival]{Surv}}, \code{\link[survival]{survfit}}}
}}

