% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ggsurvfit.R
\name{theme_ggsurvfit}
\alias{theme_ggsurvfit}
\alias{theme_ggsurvfit_default}
\alias{theme_ggsurvfit_KMunicate}
\title{Survfit Plot Themes}
\usage{
theme_ggsurvfit_default()

theme_ggsurvfit_KMunicate()
}
\value{
a ggplot2 theme
}
\description{
Returns ggplot list of calls defining a theme.
\itemize{
\item \code{theme_ggsurvfit_default()}: Builds on \code{theme_bw()} with increased text sizes.
\item \code{theme_ggsurvfit_KMunicate()}: Theme to create KMunicate-styled figures. \doi{10.1136/bmjopen-2019-030215}
}
}
\examples{
survfit2(Surv(time, status) ~ sex, data = df_lung) \%>\%
  ggsurvfit(theme = theme_ggsurvfit_default()) +
  scale_ggsurvfit()
}
